fix: c.start(tunFd) Kotlin Int 匹配 gomobile int32

+++ android/app/src/main/java/com/cp12064/moxianp2p/ClientController.kt
@@ -59,7 +59,7 @@ object ClientController {
         return try {
             appendLog("[app] launching Go client (fd=$tunFd)")
             val c = Mobile.newClient(yamlConfig, sink)
-            c.start(tunFd.toLong())
+            c.start(tunFd)
             native = c
             _running.value = true
             _state.value = State.CONNECTING
